Handle bug of Login flickering Changes to be committed: Handle bug of Login flickering 	new file:   client/src/components/LoginComp.js 	new file:   client/src/components/RegisterComp.js 	modified:   client/src/css/NewStyles.module.css 	modified:   client/src/main/App.js 	modified:   client/src/main/LoginModal.js 	modified:   client/src/main/RegisterModal.js
